Add Eips
Eips is a new, free and open source implementation of a list CRDT (for
elements of any type) written in Rust. A key feature of Eips is that all
operations run in worst-case non-amortized logarithmic time, ensuring
low latency for real-time editing applications. It also supports true
move operations that don't risk duplicating elements, and does not
suffer from interleaving issues (its correctness in this regard is
similar to Fugue).